refactored templatign of gunicorn upstart script
