--- anouman/templates/init_script/website.conf
@@ -1,5 +1,0 @@
-description     "Starting {{DOMAINNAME}}"
-
-start on startup
-exec {{GUNICORN_START}}
-respawn
